requirements.txt
aiogram==3.10.0
requests==2.31.0
python-dotenv==1.0.0
pillow==10.2.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
